moved skewering_position_sampler script to generate_cropped_images, and updated README
